Do not reset request if processing fails
If processing fails, the coroutine won't be resumed, and the death queue
will take care of the coroutine as soon as the timeout kicks in.

We still need to know if the request failed because of recoverable errors or
not; maybe the best way would be yielding directly from the place we
encountered a possibly recoverable error instead of trying to recover it
from _process_request_coro().