Add rewrite module
The pattern matching code has been shamelessly ripped from OpenBSD
http[1], which has in turn been adapted from the matching stuff from
Lua 5.3.1.  It's copied directly from OpenBSD CVS, without any changes.

The repository in [1] contains `patterns.7`, which explains the syntax.
Eventually that information will be provided as part of Lwan
documentation, but I'm too lazy to do that at this moment. In the
meantime, check that instead.

At the moment, only redirecting (and not rewriting) responses is
supported.  Rewriting requires a little bit more work but are doable.
Maybe in the future.

In addition to adding the module, the configuration file parser now
allows a section to be isolated.  Key-value pairs are passed to the
module initialization function (to the `init_from_hash()` function),
and, if the module has a `parse_conf()` function, it's called with the
isolated section as a parameter.  This allows a module to have some
default options so it's properly initialized, and then to use the state
previously initialized, and let it parse subsections as needed.  The
rewrite module does that to parse pattern subsections.

A test has been added as well.

[1] https://github.com/reyk/httpd