Revert all changes in lwan-coro since 7171a6
For some reason I couldn't identify, the redirect module when used with
Lua scripts are crashing.  The crash makes absolutely no sense (null
pointer dereference within sprintf(), while Lua tries to convert an
integer to string -- and this happens both with LuaJit and Lua.org).
I've tried a bunch of different scenarios to fix this, but nothing
worked. Reverting fixes the problem, but makes me very uneasy since
the changes are supposed to be correct.