Use a dynamically-allocated deferred callback array
when the coroutine is created.  This array has space for 16 deferred
calbacks, which is more than pentiful for all the usecases in Lwan;
more could be added later using the coro_malloc_full() primitive, if
necessary.

Ideally, this would be an array embedded within coro_t itself. However,
for some unknown reason, the Rewrite module crashes while expanding
with a Lua script.  I'll eventualy find the reason and reduce the
coroutine creation to only a single malloc() call, but for now this
should use slightly less memory and amortize the dynamic allocation
cost as the number of deferred callbacks grows.