File-backed POST body
Since there's potential for abuse, this is disabled by default. Specify
"allow_temp_files = post" in the configuration file to enable this
feature.

Once enabled, Lwan will allocate memory for the POST request bodies in
RAM if it fits under 1MiB.  If it's over (but still hardcapped to
128MiB), it'll try to create a temporary file to store the POST data.

If supported by the system (Linux >= 3.11), the temporary file will be
created using open() and O_TMPFILE.  There are many advantages in using
this, although not all filesystems supports this flag.  Please consult
the man page of open(2) for details.

If O_TMPFILE is unsupported, then mkostemps() will be used instead.
(If unavailable, a makeshift version using mkstemp()+fcntl() is
provided.)

In both cases, files will be saved in a temporary directory.  There's
currently no configuration option for this, as Lwan will try to find
the correct place by testing, in order, if the TMPDIR, TMP, and TEMP
environment variables are set.  If none of those are set, if either
/tmp or /var/tmp are directories, they'll be used instead.  (Temporary
file creation will fail if a temporary directory cannot be found.)

There's a race condition in the temp-dir-localization code, but it's
not an issue considering that the open() call within the
temp-filename-creation routine will fail, regardless if O_TMPFILE is
being used or not.

Eventually the path for temporary directories will be configurable, as
it's common practice to have /tmp mounted on a tmpfs -- which would
render this essentially useless as it wouldn't be backed by a disk.