Make ThreadSanitizer happy by declaring barrier in struct lwan
ThreadSanitizer detected a data race on the barrier as it was allocated
on the main stack but other threads used it without locking.  It seems
that it doesn't know about pthread_barriers.  Other race detectors, such
as [1], also seems to catch the exact same "issue".  Move the barrier
to the lwan struct to avoid these false positives.

[1] https://link.springer.com/article/10.1007/s11227-017-1996-8