Use readahead() call before calling sendfile()
The call is made only when an item is added to the cache, to reduce
the syscall overhead.  Since readahead() might block while filesystem
metadata is located, this is important.

It's a Linux-only call, which is stubbed on all other platforms. It's
possible to simulate it by memory-mapping the file, and:

  - If mincore(2) is available, use it to know which pages to touch.
  - If it's not available, touch all the pages.

Not sure if it's worth the trouble, though, given this is just an
optimization:

On FreeBSD, at least, Lwan uses the SF_MNOWAIT flag when calling
sendfile(); this will effectively make sendfile() return early if the
requested bytes were not in the page cache, making this optimization
useless on FreeBSD.

The macOS version of sendfile() has no such flag, and thus could
benefit from a readahead() implementation.  However, the macOS
port is considered a "best effort" port: very few people will
actually use this operating system for busy servers anyway.